Create a CAS handler.
